requirements.txt
pywin32>=300
watchdog>=2.1.6
yara-python>=4.1.3
wmi>=1.5.1
configparser>=5.2.0
flask>=2.0.0